requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
supabase==2.0.3
pandas==2.1.1
numpy==1.26.0
requests==2.31.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
MarkupSafe==2.1.3
pyyaml==6.0.1
gunicorn==21.2.0
pytest==7.4.2
pytest-cov==4.1.0
python-jose==3.3.0